+++ src/Customers/Entities/Attribute.php
@@ -6,31 +6,31 @@ use ArrowSphere\PublicApiClient\AbstractEntity;
 
 class Attribute extends AbstractEntity
 {
-    public const COLUMN_SCHEDULED_NAME = 'name';
+    public const COLUMN_NAME = 'name';
     public const COLUMN_VALUE = 'value';
 
     protected string $name;
 
-    protected string $value;
+    protected ?string $value;
 
     protected const VALIDATION_RULES = [
-        self::COLUMN_SCHEDULED_NAME => 'required|string',
-        self::COLUMN_VALUE          => 'required|string',
+        self::COLUMN_NAME  => 'required|string',
+        self::COLUMN_VALUE => 'nullable|string',
     ];
 
     public function __construct(array $data)
     {
         parent::__construct($data);
 
-        $this->value = $data['value'];
+        $this->value = $data['value'] ?? null;
         $this->name = $data['name'];
     }
 
     public function jsonSerialize(): array
     {
         return [
-            self::COLUMN_SCHEDULED_NAME => $this->name,
-            self::COLUMN_VALUE          => $this->value,
+            self::COLUMN_NAME => $this->name,
+            self::COLUMN_VALUE => $this->value,
         ];
     }
 
@@ -39,7 +39,7 @@ class Attribute extends AbstractEntity
         return $this->name;
     }
 
-    public function getValue(): string
+    public function getValue(): ?string
     {
         return $this->value;
     }